packages/plugin-autogen/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-autogen"
version = "0.1.0"
description = "Microsoft AutoGen integrations for TinyHumans memory API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "tinyhumansai>=0.5",
    "pyautogen>=0.2.20",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
examples = ["python-dotenv>=1.0"]

[tool.setuptools.packages.find]
include = ["neocortex_autogen*"]

[tool.ruff]
line-length = 88
target-version = "py39"
